requirements.txt
numpy
pyyaml
sklearn
matplotlib
prefetch_generator
wandb
torch
torchvision
tensorboard
hydra-core
hydra_colorlog 
omegaconf
pip install git+https://github.com/ildoonet/pytorch-randaugment
